--- fltk/examples/pixmap.rs
@@ -1,61 +1,0 @@
-use fltk::*;
-
-const PXM: &[&str] = &[
-    "50 34 4 1",
-    "  c #000000",
-    "o c #ff9900",
-    "@ c #ffffff",
-    "# c None",
-    "##################################################",
-    "###      ##############################       ####",
-    "### ooooo  ###########################  ooooo ####",
-    "### oo  oo  #########################  oo  oo ####",
-    "### oo   oo  #######################  oo   oo ####",
-    "### oo    oo  #####################  oo    oo ####",
-    "### oo     oo  ###################  oo     oo ####",
-    "### oo      oo                     oo      oo ####",
-    "### oo       oo  ooooooooooooooo  oo       oo ####",
-    "### oo        ooooooooooooooooooooo        oo ####",
-    "### oo     ooooooooooooooooooooooooooo    ooo ####",
-    "#### oo   ooooooo ooooooooooooo ooooooo   oo #####",
-    "####  oo oooooooo ooooooooooooo oooooooo oo  #####",
-    "##### oo oooooooo ooooooooooooo oooooooo oo ######",
-    "#####  o ooooooooooooooooooooooooooooooo o  ######",
-    "###### ooooooooooooooooooooooooooooooooooo #######",
-    "##### ooooooooo     ooooooooo     ooooooooo ######",
-    "##### oooooooo  @@@  ooooooo  @@@  oooooooo ######",
-    "##### oooooooo @@@@@ ooooooo @@@@@ oooooooo ######",
-    "##### oooooooo @@@@@ ooooooo @@@@@ oooooooo ######",
-    "##### oooooooo  @@@  ooooooo  @@@  oooooooo ######",
-    "##### ooooooooo     ooooooooo     ooooooooo ######",
-    "###### oooooooooooooo       oooooooooooooo #######",
-    "###### oooooooo@@@@@@@     @@@@@@@oooooooo #######",
-    "###### ooooooo@@@@@@@@@   @@@@@@@@@ooooooo #######",
-    "####### ooooo@@@@@@@@@@@ @@@@@@@@@@@ooooo ########",
-    "######### oo@@@@@@@@@@@@ @@@@@@@@@@@@oo ##########",
-    "########## o@@@@@@ @@@@@ @@@@@ @@@@@@o ###########",
-    "########### @@@@@@@     @     @@@@@@@ ############",
-    "############  @@@@@@@@@@@@@@@@@@@@@  #############",
-    "##############  @@@@@@@@@@@@@@@@@  ###############",
-    "################    @@@@@@@@@    #################",
-    "####################         #####################",
-    "##################################################",
-];
-
-fn main() {
-    let app = app::App::default();
-    app::set_visual(Mode::Rgb8).unwrap();
-    let mut win = window::Window::default().with_size(500, 400);
-    let mut frame = frame::Frame::new(0, 0, 500, 400, "");
-    win.end();
-    win.make_resizable(true);
-    win.show();
-
-    frame.draw(|| {
-        let mut pxm = image::Pixmap::new(PXM);
-        pxm.scale(500, 400, false, true);
-        pxm.draw(0, 0, 500, 400);
-    });
-
-    app.run().unwrap();
-}
