--- fltk/examples/timeout.rs
@@ -1,18 +1,0 @@
-// Calque example of the FLTK docs
-
-use fltk::*;
-
-fn callback() {
-    println!("TICK");
-    app::repeat_timeout(1.0, callback);
-}
-
-fn main() {
-    let app = app::App::default();
-    let mut wind = window::Window::new(100, 100, 400, 300, "");
-    wind.show();
-
-    app::add_timeout(1.0, callback);
-
-    app.run().unwrap();
-}
