+++ fltk/src/surface.rs
@@ -4,6 +4,17 @@ use std::ffi::CString;
 use std::path;
 
 /// An image surface object
+/// Example usage:
+/// ```no_run
+/// let but = button::Button::new(0, 0, 80, 40, "Click");
+/// let sur = surface::ImageSurface::new(but.width(), but.height(), false);
+/// surface::ImageSurface::push_current(&sur);
+/// draw::set_draw_color(Color::White);
+/// draw::draw_rectf(0, 0, but.width(), but.height());
+/// sur.draw(&but, 0, 0);
+/// let img = sur.image().unwrap();
+/// surface::ImageSurface::pop_current();
+/// ```
 pub struct ImageSurface {
     _inner: *mut Fl_Image_Surface,
 }
@@ -119,7 +130,20 @@ impl Drop for ImageSurface {
     }
 }
 
-/// An image surface object
+/// An SVG image surface object
+/// Example usage:
+/// ```no_run
+/// let but = button::Button::new(0, 0, 80, 40, "Click");
+/// // We need the destructor of SvgFileSurface to actually create the image
+/// {
+///     let sur = surface::SvgFileSurface::new(but.width(), but.height(), "temp.svg");
+///     surface::SvgFileSurface::push_current(&sur);
+///     draw::set_draw_color(Color::White);
+///     draw::draw_rectf(0, 0, but.width(), but.height());
+///     sur.draw(&but, 0, 0);
+///     surface::SvgFileSurface::pop_current();
+/// }
+/// ```
 pub struct SvgFileSurface {
     _inner: *mut Fl_SVG_File_Surface,
 }
